Re-arrange namespaces.
