+++ framework/src/Volo.Abp.BlobStoring/Volo/Abp/BlobStoring/AbpBlobStoringModule.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Volo.Abp.BlobStoring.Containers;
 using Volo.Abp.Modularity;
 
 namespace Volo.Abp.BlobStoring
