+++ framework/src/Volo.Abp.BlobStoring/Volo/Abp/BlobStoring/BlobProviderDeleteArgs.cs
@@ -1,12 +1,11 @@
 ﻿using System.Threading;
 using JetBrains.Annotations;
-using Volo.Abp.BlobStoring.Containers;
 
-namespace Volo.Abp.BlobStoring.Providers
+namespace Volo.Abp.BlobStoring
 {
-    public class BlobProviderGetArgs : BlobProviderArgs
+    public class BlobProviderDeleteArgs : BlobProviderArgs
     {
-        public BlobProviderGetArgs(
+        public BlobProviderDeleteArgs(
             [NotNull] string containerName,
             [NotNull] BlobContainerConfiguration configuration,
             [NotNull] string blobName,
