+++ framework/src/Volo.Abp.BlobStoring/Volo/Abp/BlobStoring/BlobProviderExistsArgs.cs
@@ -1,8 +1,7 @@
 ﻿using System.Threading;
 using JetBrains.Annotations;
-using Volo.Abp.BlobStoring.Containers;
 
-namespace Volo.Abp.BlobStoring.Providers
+namespace Volo.Abp.BlobStoring
 {
     public class BlobProviderExistsArgs : BlobProviderArgs
     {
