--- framework/test/Volo.Abp.BlobStoring.Tests/Volo/Abp/BlobStoring/Containers/TestObjects/TestContainer3.cs
@@ -1,7 +1,0 @@
-﻿namespace Volo.Abp.BlobStoring.Containers.TestObjects
-{
-    public class TestContainer3
-    {
-        
-    }
-}
