+++ framework/test/Volo.Abp.BlobStoring.Tests/Volo/Abp/BlobStoring/TestObjects/TestContainer1.cs
@@ -1,0 +1,7 @@
+﻿namespace Volo.Abp.BlobStoring.TestObjects
+{
+    public class TestContainer1
+    {
+        
+    }
+}
